+++ py-polars/src/expr.rs
@@ -23,3 +23,29 @@ impl From<Expr> for PyExpr {
         PyExpr { inner: expr }
     }
 }
+
+pub(crate) trait ToExprs {
+    fn to_exprs(self) -> Vec<Expr>;
+}
+
+impl ToExprs for Vec<PyExpr> {
+    fn to_exprs(self) -> Vec<Expr> {
+        // Safety
+        // repr is transparent
+        // and has only got one inner field`
+        unsafe { std::mem::transmute(self) }
+    }
+}
+
+pub(crate) trait ToPyExprs {
+    fn to_pyexprs(self) -> Vec<PyExpr>;
+}
+
+impl ToPyExprs for Vec<Expr> {
+    fn to_pyexprs(self) -> Vec<PyExpr> {
+        // Safety
+        // repr is transparent
+        // and has only got one inner field`
+        unsafe { std::mem::transmute(self) }
+    }
+}
